+++ src/providers/index.ts
@@ -1,3 +1,4 @@
+export * from './LoadingOverlayProvider';
 export * from './PrimeReactProvider';
 export * from './Providers';